Transform upgrades/driver.py into package
Kostyor is going to have multiple upgrade drivers and having them in one
module is not convenient from maintaining and developing point of view.
So let's transform this module into a package, so each new driver could
be encapsulated inside a separate module.
