+++ kostyor/upgrades/__init__.py
@@ -1,9 +1,6 @@
 from kostyor.upgrades.engine import Engine
-from kostyor.upgrades.driver import UpgradeDriver, NoOpDriver
 
 
 __all__ = [
     'Engine',
-    'UpgradeDriver',
-    'NoOpDriver',
 ]
