+++ setup.cfg
@@ -1,13 +1,13 @@
 [metadata]
 name = kostyor
 summary = upgrade tool
-description-file = 
+description-file =
     README.rst
 author = Sean M. Collins
 author-email = sean@coreitpro.com
 
 [files]
-packages = 
+packages =
     kostyor
 
 [build_sphinx]
@@ -23,7 +23,8 @@ warnerrors = true
 
 
 [entry_points]
-kostyor.upgrade_drivers =
-    noop = kostyor.upgrades.driver:NoOpDriver
+kostyor.upgrades.drivers =
+    noop = kostyor.upgrades.drivers.noop:NoOpDriver
+
 kostyor.discovery_drivers =
     openstack = kostyor.inventory.discover:OpenStackServiceDiscovery
